fix: break node exec host helper import cycle
