+++ src/agents/bash-tools.exec-host-node.types.ts
@@ -1,0 +1,27 @@
+import type { ExecAsk, ExecSecurity } from "../infra/exec-approvals.js";
+
+export type ExecuteNodeHostCommandParams = {
+  command: string;
+  workdir: string | undefined;
+  env: Record<string, string>;
+  requestedEnv?: Record<string, string>;
+  requestedNode?: string;
+  boundNode?: string;
+  sessionKey?: string;
+  turnSourceChannel?: string;
+  turnSourceTo?: string;
+  turnSourceAccountId?: string;
+  turnSourceThreadId?: string | number;
+  trigger?: string;
+  agentId?: string;
+  security: ExecSecurity;
+  ask: ExecAsk;
+  strictInlineEval?: boolean;
+  timeoutSec?: number;
+  defaultTimeoutSec: number;
+  approvalRunningNoticeMs: number;
+  warnings: string[];
+  notifySessionKey?: string;
+  notifyOnExit?: boolean;
+  trustedSafeBinDirs?: ReadonlySet<string>;
+};
